requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
pydub==0.25.1
pyTelegramBotAPI==4.18.1
python-dotenv==1.0.1
requests==2.32.3
SpeechRecognition==3.10.4
telebot==0.0.5
typing_extensions==4.12.0
urllib3==2.2.1
